chore: move init history logic to service
