+++ packages/core/content-manager/server/src/history/utils.ts
@@ -1,3 +1,4 @@
+import { Strapi } from '@strapi/types';
 import { type services } from './services';
 
 type HistoryServiceName = keyof typeof import('./services').services;
@@ -6,7 +7,7 @@ type HistoryServices = {
   [K in HistoryServiceName]: ReturnType<(typeof services)[K]>;
 };
 
-function getService<T extends keyof HistoryServices>(name: T) {
+function getService<T extends keyof HistoryServices>(strapi: Strapi, name: T) {
   // Cast is needed because the return type of strapi.service is too vague
   return strapi.service(`plugin::content-manager.${name}`) as HistoryServices[T];
 }
